Delete saved data and re-run
To ensure recent changes didn't break the REST API logic.

The new data are saved. The goal here is not to stop tracking the
saved data, but to ensure recent major changes didn't break stuff.

(The data is a little bit different now, too, but that's not the
reason I'm re-retrieving it.)