Have the logger build the strings
So they don't have to be built if the output is not used.

Note that the significant change is here is in stars.py. The change
in stars.ipynb is just because I didn't delete the saved file, so
it loaded it from the cache. Currently this project doesn't use a
GitHub API key, so rate limiting is very strict I can't re-run the
API request very frequently.